Restuarnet menu code moved to utilities file.
